+++ src/Bench.php
@@ -2,6 +2,8 @@
 
 namespace ModernPHPException;
 
+use LogicException;
+
 class Bench
 {
     /**
@@ -37,9 +39,7 @@ class Bench
      */
     public function end(): void
     {
-        if (!$this->hasStarted()) {
-            throw new \LogicException("You must call start()");
-        }
+        if (!$this->hasStarted()) throw new LogicException("You must call start()");
 
         $this->end_time = microtime(true);
         $this->memory_usage = memory_get_usage(true);
@@ -56,16 +56,10 @@ class Bench
      */
     public function getTime(bool $raw = false, string $format = null): mixed
     {
-        if (!$this->hasStarted()) {
-            throw new \LogicException("You must call start()");
-        }
-
-        if (!$this->hasEnded()) {
-            throw new \LogicException("You must call end()");
-        }
+        if (!$this->hasStarted()) throw new LogicException("You must call start()");
+        if (!$this->hasEnded()) throw new LogicException("You must call end()");
 
         $elapsed = $this->end_time - $this->start_time;
-
         return $raw ? $elapsed : self::readableElapsedTime($elapsed, $format);
     }
 
@@ -93,7 +87,6 @@ class Bench
     public function getMemoryPeak(bool $raw = false, ?string $format = null): mixed
     {
         $memory = memory_get_peak_usage(true);
-
         return $raw ? $memory : self::readableSize($memory, $format);
     }
 
@@ -131,21 +124,14 @@ class Bench
     public static function readableSize(int $size, ?string $format = null, int $round = 3): string
     {
         $mod = 1024;
-
-        if (is_null($format)) {
-            $format = '%.2f%s';
-        }
-
+        if (is_null($format)) $format = '%.2f%s';
         $units = explode(' ', 'B Kb Mb Gb Tb');
 
         for ($i = 0; $size > $mod; $i++) {
             $size /= $mod;
         }
 
-        if (0 === $i) {
-            $format = preg_replace('/(%.[\d]+f)/', '%d', $format);
-        }
-
+        if (0 === $i) $format = preg_replace('/(%.[\d]+f)/', '%d', $format);
         return sprintf($format, round($size, $round), $units[$i]);
     }
 
@@ -160,9 +146,7 @@ class Bench
      */
     public static function readableElapsedTime(float $microtime, ?string $format = null, int $round = 3): string
     {
-        if (is_null($format)) {
-            $format = '%.3f%s';
-        }
+        if (is_null($format)) $format = '%.3f%s';
 
         if ($microtime >= 1) {
             $unit = 's';